+++ mobile/native/migrations/2023-08-07-232900_add_order_expiry_to_order/down.sql
@@ -1,2 +1,2 @@
 ALTER TABLE
-    orders DROP COLUMN IF EXISTS order_expiry_timestamp;
+    orders DROP COLUMN order_expiry_timestamp;