+++ mobile/native/migrations/2023-10-25-120015_add_fee_to_channel/up.sql
@@ -1,0 +1,5 @@
+-- Your SQL goes here
+ALTER TABLE "channels" ADD COLUMN "fee_sats" BIGINT;
+ALTER TABLE "channels" ADD COLUMN "open_channel_payment_hash" TEXT;
+
+ALTER TABLE "payments" ADD COLUMN "funding_txid" TEXT;